rewriting entity selection logic
so that it's like a proper rts, and not shit.
this commit is being committed so that i can make a commit where I change one thing, and that is the selectedEntities vector. i think it needs to be a reference list to things in the entity list vector instead of a seperate list itself.
i think this causes some problems with entities thinking they're selected if they really aren't. we will see.