added a Camera class which is a wrapper for sf::View
but was it really necesarry? I'm kind of regretting it thinking I could have kept what I had but I feel ilke in the long run this might pay off. I guess we'll see later on but for now I'll keep it.

also tried to optimise some of the EntityManager. not sure how well it went but deleting entities is noticably faster. there is a memory leak somewhere in there, however. im not sure where it is.