Interface rework: it's much nicer now
the bottom bar has been introduced. it's not completely functional, but it should be soon. this also comes with a bigger window, 720p. only issue ive found with this is that the debug info on selected bobs is blurry when below default zoom level
entity manager cleans up selected ents on destruction, too
app engine frees all textures when cleaning up