added commentsection entity
it'll be like the barracks in age of empires. it's currently non function, but the next commit should be the one that adds the proper functionality. or it'll be the one that updates the interface system to actually support entities that makes stuff like that, because i don't think it'll work well in it's current form.