pick top entities when selecting an entity
because i reveresed the entity iteration loop
also googleplus now move to random location when they spawn. this is just to give us some stuff to look at for now. eventually, they will not be moveable by the player and will spawn next to their googles or whatever it is that makes them.